Associate users and stories with many to many relationship
user_stories is a join table to support a many to many rlationship. Although the MVP of this project does not use user.stories, it is a useful association for the future and doesn't complicate the code. Including the user_stories table now prevents us from having to perform migrations to create this association laster in development if we only had a one-to-many relationship at this stage.